Backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
opencv-python==4.8.1.78
pillow==10.1.0
numpy~=1.26.4
tensorflow==2.16.1
scikit-learn==1.3.2
websockets==12.0
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
pydantic-settings==2.1.0
aiofiles==23.2.1
websockets==12.0
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
psutil>=5.9.0
